backend/pyproject.toml
[project]
name = "arxi"
version = "0.1.0"
description = "AI-native pharmacy management system"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "redis>=5.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "defusedxml>=0.7.0",
    "pytesseract>=0.3.13",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
